# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Nestor Vocal API
  description: >
    API REST du concierge vocal intelligent Nestor Vocal — Selectour Alltour.
    Deux systèmes d'authentification JWT distincts : un pour les utilisateurs
    (clients de l'agence), un pour les administrateurs.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Serveur local de développement
...
components:
  securitySchemes:
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT utilisateur, obtenu via /api/auth/login ou /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT administrateur, obtenu via /api/admin/auth/login (secret séparé du JWT utilisateur)
  schemas:
...
    Appointment:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        title: { type: string }
        description: { type: string }
        dateTime: { type: string, format: date-time }
        agent: { type: string }
        quantity: { type: integer, minimum: 1 }
        location: { type: string }
        status: { type: string, enum: [upcoming, completed, cancelled] }
    WaitlistEntry:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        date: { type: string, format: date }
        quantity: { type: integer }
        rank: { type: integer }
        total: { type: integer }
...
security:
  - UserAuth: []
paths:
...
  /api/appointments:
    get:
      tags: [Appointments]
      summary: Liste des rendez-vous de l'utilisateur connecté
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Appointment' } } } }
    post:
      tags: [Appointments]
      summary: Créer un rendez-vous
      description: Vérifie la disponibilité du créneau avant création (assertSlotAvailable).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, dateTime]
              properties:
                title: { type: string }
                description: { type: string }
                dateTime: { type: string, format: date-time }
                quantity: { type: integer, minimum: 1 }
                agent: { type: string }
      responses:
        '201': { description: Créé, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '409': { description: Créneau déjà réservé (SLOT_TAKEN) }
  /api/appointments/taken:
    get:
      tags: [Appointments]
      summary: Créneaux déjà pris pour une date donnée
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: "Liste des horaires occupés (HH:MM)" }
  /api/appointments/{id}:
    put:
      tags: [Appointments]
      summary: Modifier un rendez-vous
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } }
      responses:
        '200': { description: Rendez-vous mis à jour }
    delete:
      tags: [Appointments]
      summary: Supprimer un rendez-vous
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Supprimé }
  /api/appointments/waitlist:
    get:
      tags: [Appointments]
      summary: Liste d'attente pour une date donnée
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses:
        '200': { content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/WaitlistEntry' } } } } }
    post:
      tags: [Appointments]
      summary: S'inscrire en liste d'attente
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [date, name]
              properties:
                date: { type: string, format: date }
                name: { type: string }
                quantity: { type: integer }
      responses:
        '201': { description: Inscrit, content: { application/json: { schema: { $ref: '#/components/schemas/WaitlistEntry' } } } }
        '409': { description: "Déjà inscrit ce jour (ALREADY_LISTED)" }
  /api/appointments/waitlist/me:
    get:
      tags: [Appointments]
      summary: Ma position dans la liste d'attente pour une date
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/WaitlistEntry' } } } }
  /api/appointments/waitlist/{id}:
    delete:
      tags: [Appointments]
      summary: Se désinscrire de la liste d'attente
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Désinscrit }

  # ─────────────────────────── TICKETS ───────────────────────────
...
  /api/admin/appointments:
    get:
      tags: [Admin]
      summary: Liste des rendez-vous (tous utilisateurs, filtrable)
      security: [{ AdminAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200': { description: OK }
  /api/admin/appointments/{id}:
    put:
      tags: [Admin]
      summary: Modifier un rendez-vous (admin)
      security: [{ AdminAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Mis à jour }
    delete:
      tags: [Admin]
      summary: Supprimer un rendez-vous (admin)
      security: [{ AdminAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Supprimé }
...